move new-book page
