--- apps/client/src/features/create-book/index.ts
@@ -1,2 +1,0 @@
-/** @public */
-export { CreateBookForm } from "./ui";
