--- apps/client/src/features/create-book/ui/index.ts
@@ -1,1 +1,0 @@
-export { CreateBookForm } from "./new-book-form";
