+++ apps/client/src/pages/new-book/api/create-book.graphql
