+++ apps/client/src/pages/new-book/index.ts
@@ -1,0 +1,2 @@
+/** @public */
+export { NewBookPage } from "./ui";
