+++ apps/client/src/pages/new-book/ui/index.ts
@@ -1,0 +1,1 @@
+export { NewBookPage } from "./page";
