+++ apps/client/src/routes/books/new.tsx
@@ -1,28 +1,8 @@
-import { Link, createFileRoute, useNavigate } from "@tanstack/react-router";
-import { BiLeftArrowAlt } from "react-icons/bi";
-import type { BookRef } from "~/entities/book";
-import { CreateBookForm } from "~/features/create-book";
+import { createFileRoute } from "@tanstack/react-router";
+import { NewBookPage } from "~/pages/new-book";
 
 const RouteComponent: React.FC = () => {
-  const navigate = useNavigate();
-
-  const onCreatedBook = (bookRef: BookRef) => {
-    void navigate({
-      to: "/books/$book-ref",
-      params: { "book-ref": bookRef },
-    });
-  };
-
-  return (
-    <div className=":uno: px-8 py-4">
-      <Link to="/books" className=":uno: color-inherit decoration-none">
-        <BiLeftArrowAlt />
-        <span className=":uno: ml-1 align-middle">Return to books</span>
-      </Link>
-      <hr className=":uno: my-2 border-gray-7 border-none border-t-solid" />
-      <CreateBookForm onCreatedBook={onCreatedBook} />
-    </div>
-  );
+  return <NewBookPage />;
 };
 
 export const Route = /* #__PURE__ */ createFileRoute("/books/new")({
